Refactoring is done, tests are added
