--- hw1_4/hw1_4.sln
@@ -1,17 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "hw1_4", "hw1_4\hw1_4.csproj", "{CA251F14-8CA6-42CD-A3C0-398AF32DE265}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x86 = Debug|x86
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CA251F14-8CA6-42CD-A3C0-398AF32DE265}.Debug|x86.ActiveCfg = Debug|x86
-		{CA251F14-8CA6-42CD-A3C0-398AF32DE265}.Debug|x86.Build.0 = Debug|x86
-		{CA251F14-8CA6-42CD-A3C0-398AF32DE265}.Release|x86.ActiveCfg = Release|x86
-		{CA251F14-8CA6-42CD-A3C0-398AF32DE265}.Release|x86.Build.0 = Release|x86
-	EndGlobalSection
-EndGlobal
