--- hw2_1/hw2_1.userprefs
@@ -1,12 +1,0 @@
-﻿<Properties StartupItem="hw2_1/hw2_1.csproj">
-  <MonoDevelop.Ide.Workspace ActiveConfiguration="Debug|x86" />
-  <MonoDevelop.Ide.Workbench ActiveDocument="hw2_1/Program.cs">
-    <Files>
-      <File FileName="hw2_1/Program.cs" Line="34" Column="18" />
-    </Files>
-  </MonoDevelop.Ide.Workbench>
-  <MonoDevelop.Ide.DebuggingService.Breakpoints>
-    <BreakpointStore />
-  </MonoDevelop.Ide.DebuggingService.Breakpoints>
-  <MonoDevelop.Ide.DebuggingService.PinnedWatches />
-</Properties>
