+++ hw2_1/hw2_1/Program.cs
@@ -2,61 +2,7 @@
 
 namespace ArrayStack
 {
-	class ArrayStack
-	{
-		private const int maxElementsNumber = 100; // max number of elements in the stack array
-		private int currentSize = 0; // number of elements in the stack array
-		private int[] array = new int[maxElementsNumber]; 
-
-		// class constructor
-		public ArrayStack() 
-		{
-			for (int i = 0; i < maxElementsNumber; ++i)
-			{
-				this.array[i] = 0; // just in case
-			}
-			currentSize = 0; // just in case
-		}
-
-		// returns true if stack is empty, otherwise it's false
-		public bool isEmpty() 
-		{
-			if (this.currentSize == 0) 
-			{
-				return true;
-			} 
-			else 
-			{
-				return false;
-			}	
-		}
-
-		// returns true if element is added, and false if stack is going to get overloaded
-		public bool Push(int element)
-		{
-			if (currentSize < maxElementsNumber) 
-			{
-				this.array[currentSize] = element;
-				++currentSize;
-				return true;
-			} 
-			else 
-			{
-				return false;
-			}
-		}
-
-		// returns the last added element, deletes it from the stack
-		// warning: one should check if the stack is not empty before this operation
-		public int Pop()
-		{
-			--currentSize;
-			return this.array[currentSize]; 
-		}
-
-	}
-
-	class MainClass
+	class Program
 	{
 		public static void Main (string[] args)
 		{
